+++ matrix2d.h
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include <time.h>
 
-int **generate_matrix(int rows, int cols);
-void rand_adj_matrix(int **matrix, int size);
-void print_matrix(int **matrix, int rows, int cols);
+int **generate_matrix(long rows, long cols);
+void rand_adj_matrix(int **matrix, long size);
+void print_matrix(int **matrix, long rows, long cols);
 void free_matrix(int **matrix);
